Player design done, home page modified and organization modified
