+++ lib/spotify_interface_web/live/components/player_bar/player_bar_component.ex
@@ -1,0 +1,8 @@
+defmodule SpotifyInterfaceWeb.PlayerBarComponent do
+  use SpotifyInterfaceWeb, :live_component
+  alias SpotifyInterface.Services.SpotifyService
+
+  def mount(params, session, socket) do
+    {:ok, socket}
+  end
+end
