+++ lib/spotify_interface_web/live/components/player_bar/player_bar_component.html.heex
@@ -1,0 +1,73 @@
+<div id="player" class="fixed w-full min-w-[300px] bottom-0 h-24 flex flex-row bg-black bg-clip-padding backdrop-filter backdrop-blur-sm bg-opacity-20">
+    <div id="cover-and-title" class="w-2/3 sm:w-1/3 flex items-center gap-x-2 px-4">
+        <div id="cover" class="w-16 md:w-24 h-16 md:h-24 py-2">
+            <img src="https://i.scdn.co/image/ab67616d0000b27349d694203245f241a1bcaa72" class="object-cover h-full w-full" />
+        </div>
+        <div id="title" class="w-8/12 h-24 text-gray-200 flex flex-col justify-center py-2">
+            <h2 class="font-bold">Me porto bonito</h2>
+            <div class="text-gray-300 overflow-ellipsis whitespace-nowrap overflow-x-hidden capitalize">Bad Bunny</div>
+        </div>
+    </div>
+
+    <div id="controlers" class="w-1/3 hidden sm:flex flex-col h-full">
+        <div class="h-3/5 flex justify-center items-center gap-x-2">
+            <div class="rounded-full h-8 w-8 text-gray-300 hover:text-gray-100 cursor-pointer flex">
+                <svg class="p-2 fill-current" viewBox="0 0 32 32">
+                    <g>
+                        <path d="M30.5,2.62a1,1,0,0,0-1,0L7.82,15.13a1,1,0,0,0,0,1.74L29.5,29.38a1,1,0,0,0,1.5-.86v-25A1,1,0,0,0,30.5,2.62ZM29,26.78,10.32,16,29,5.22Z" />
+                        <path d="M2,5.2a1,1,0,0,0-1,1V25.8a1,1,0,0,0,2,0V6.2A1,1,0,0,0,2,5.2Z" />
+                    </g>
+                </svg>
+            </div>
+            <div class="rounded-full h-10 w-10 text-gray-100 cursor-pointer flex">
+                <i class="hero-play h-10 w-10 p-2"></i>
+            </div>
+            <div class="rounded-full h-8 w-8 text-gray-300 hover:text-gray-100 cursor-pointer flex">
+                <svg class="p-2 fill-current" viewBox="0 0 32 32">
+                    <g>
+                        <path d="M24.18,15.13,2.5,2.62A1,1,0,0,0,1,3.48v25a1,1,0,0,0,1.5.86L24.18,16.87A1,1,0,0,0,24.18,15.13ZM3,26.78V5.22L21.68,16Z" />
+                        <path d="M30,5.2a1,1,0,0,0-1,1V25.8a1,1,0,0,0,2,0V6.2A1,1,0,0,0,30,5.2Z" />
+                    </g>
+                </svg>
+            </div>
+        </div>
+        <div class="h-2/5 w-full flex justify-between items-center">
+            <span class="text-sm text-gray-200 w-12 text-left">1:20</span>
+            <input type="range" class="w-full h-1" />
+            <span class="text-sm text-gray-200 w-12 text-right">4:16</span>
+        </div>
+    </div>
+
+    <div id="controlers-mobile" class="w-1/3 flex sm:hidden flex-col h-full">
+        <div class="h-full flex justify-center items-center gap-x-1">
+            <div class="rounded-full h-8 w-8 text-gray-300 hover:text-gray-100 cursor-pointer flex">
+                <svg class="p-2 fill-current" viewBox="0 0 32 32">
+                    <g>
+                        <path d="M30.5,2.62a1,1,0,0,0-1,0L7.82,15.13a1,1,0,0,0,0,1.74L29.5,29.38a1,1,0,0,0,1.5-.86v-25A1,1,0,0,0,30.5,2.62ZM29,26.78,10.32,16,29,5.22Z" />
+                        <path d="M2,5.2a1,1,0,0,0-1,1V25.8a1,1,0,0,0,2,0V6.2A1,1,0,0,0,2,5.2Z" />
+                    </g>
+                </svg>
+            </div>
+            <div class="rounded-full h-10 w-10 text-gray-100 cursor-pointer flex justify-center items-center">
+                <i class="hero-play h-10 w-10 p-2"></i>
+            </div>
+            <div class="rounded-full h-8 w-8 text-gray-300 hover:text-gray-100 cursor-pointer flex">
+                <svg class="p-2 fill-current" viewBox="0 0 32 32">
+                    <g>
+                        <path d="M24.18,15.13,2.5,2.62A1,1,0,0,0,1,3.48v25a1,1,0,0,0,1.5.86L24.18,16.87A1,1,0,0,0,24.18,15.13ZM3,26.78V5.22L21.68,16Z" />
+                        <path d="M30,5.2a1,1,0,0,0-1,1V25.8a1,1,0,0,0,2,0V6.2A1,1,0,0,0,30,5.2Z" />
+                    </g>
+                </svg>
+            </div>
+        </div>
+    </div>
+
+    <div id="volume" class="w-1/3 hidden sm:flex justify-end items-center text-gray-100 px-4">
+        <div class="w-5 flex justify-center items-center mx-1">
+            <i class="hero-speaker-wave h-5 w-5"></i>
+        </div>
+        <div class="w-24 flex justify-center items-center">
+            <input class="h-1 w-full" type="range" />
+        </div>
+    </div>
+</div>
